Rename namespace methods to reflect the fact that they handle category rather than namespace propagation
