+++ lib/get-runtime-data.js
@@ -4,7 +4,7 @@
 
 const loadJSON = require('./load-json')
 const mergeInstances = require('./merge-instances')
-const propagateNamespace = require('./propagate-namespace')
+const propagateNamespace = require('./propagate-categories')
 const propagateSteps = require('./propagate-steps')
 const propagateShow = require('./propagate-show')
 const injectRepeatablePages = require('./inject-repeatable-pages')
@@ -15,8 +15,8 @@ const getRuntimeData = (sourceObjs, schemas) => {
     .then(sourceInstances => {
       const mergedInstances = mergeInstances.merge(sourceInstances)
       const stepsInstances = propagateSteps.propagate(mergedInstances)
-      const namespaceInstances = propagateNamespace.propagate(stepsInstances, schemas)
-      const showInstances = propagateShow.propagate(namespaceInstances)
+      const categoryInstances = propagateNamespace.propagate(stepsInstances, schemas)
+      const showInstances = propagateShow.propagate(categoryInstances)
       const repeatableInstances = injectRepeatablePages.inject(showInstances)
       const i18nRouteInstances = internationalizeRoutes.i18nUrls(repeatableInstances)
       const instances = i18nRouteInstances
