+++ lib/get-runtime-data.unit.spec.js
@@ -18,7 +18,7 @@ test('When transforming the edit-time instances into run-time ones', function (t
   t.plan(2)
 
   const sourceObjs = [source1]
-  const schemas = getTestData('namespace/schemas')
+  const schemas = getTestData('categories/schemas')
   const expected = getTestData('runtime/expected-a')
   getRuntimeData(sourceObjs, schemas)
     .then(instances => {
