+++ lib/propagate-categories.js
@@ -1,12 +1,15 @@
 /**
- * @module propagateNamespace
+ * @module propagateCategories
  **/
+
+//  NB. this propagation method could be deprecated
+// this info could arguably be better obtained by providing a helper method to query an instance’s schema
 
 const jp = require('jsonpath')
 const {deepClone} = require('@ministryofjustice/fb-utils-node')
 
 /**
- * Propagate namespace information through nested instances
+ * Propagate category information through nested instances
  *
  * @param {object} instances
  *  Object of service instances keyed by id
@@ -15,7 +18,7 @@ const {deepClone} = require('@ministryofjustice/fb-utils-node')
  *  Object of service schemas keyed by _name
  *
  * @return {object}
- *  Cloned object containing instances with propagated namespaces
+ *  Cloned object containing instances with propagated categorys
  **/
 const propagate = (instances, schemas) => {
   instances = deepClone(instances)
